data_pipeline/requirements.txt
requests
pandas
google-cloud-bigquery
python-dotenv
